Store and keep updated state rather than reading from the elements
Numerous problems with the other approach, one big one is that the id
will have restrictions as it's set to the class

Only advantages were that state would never have to worry about being
out of sync and could get elements by id using selectors. Both are less
of an advantage then the disadvantages of the method.